refactor(runtime): retain package file baseline
